Size, tests, and various fixes (#41)
* reset

* scaffold

* baseHooks and riskConfigs

* types

* credit token

* locker

* open & close

* constructor

* stage & exchange

* deposit

* withdraw

* tweak params ordering

* bug fix

* nonFungible

* oracle & interest

* debt & seize

* hooks & extsload

* size reduce

* Fix position and debt token (#38)

* 04 locker test (#29)

* Add risk config test

* Add credit tokens test

* Add fungible test

* Add Locker test

* Add more locker test

* Fix test

* Add position test

* Add fungible and licredity base test

* Add non fungible deposit test

* Add basic test

* Fix events in Licredity

* Add extsload in ILicredity

* Add debt increase and decrease test

* Add close position test

* Fix position and debt token

* comments & tweaks

* fix return

* tweak

* tweak

* typo

* disburseInterest(accrueOnly) -> collectInterest(distribute)

* receive, exactOut, defensive interest rate, and seize event

* accrue only if no liquidity present

* Licredity test (#40)

* 04 locker test (#29)

* Add risk config test

* Add credit tokens test

* Add fungible test

* Add Locker test

* Add more locker test

* Fix test

* Add position test

* Add fungible and licredity base test

* Add non fungible deposit test

* Add basic test

* Fix events in Licredity

* Add extsload in ILicredity

* Add debt increase and decrease test

* Add close position test

* Fix position and debt token

* Add debt token test

* Delete safe cast

* Fix risk config test

* Add position library test

* Fix risk mock test

* Add licredity position test

* Rename Debt token in test

* Add interest rate test

* Fix position test

* Delete statelibrary test

* Add position seize test

* Fix test after merge

* Add hook test

* Fix seize event

* Add hooks test

---------


---------

+++ test/libraries/FungibleState.t.sol
@@ -5,7 +5,7 @@ import {Test} from "@forge-std/Test.sol";
 import {FungibleState, toFungibleState} from "src/types/FungibleState.sol";
 
 contract FungibleStateTest is Test {
-    function test_FungibleState_packAndUnpack(uint64 index, uint192 balance) public pure {
+    function test_FungibleState_packAndUnpack(uint64 index, uint128 balance) public pure {
         FungibleState state = toFungibleState(index, balance);
         assertEq(state.index(), index);
         assertEq(state.balance(), balance);
